Fish Display Working and Return to Boat Changes
When a minigame is won or lost it will display a message indicating such.

To return to the boat the conditions has been changed from Max Capacity of fish being reached to Attempts made. Attempts are increased whenever the minigame is started

When return to boat happens, the fish caught during the descent is shown with its metadata